Refactorización
separación de la pagina en componentes individuales
